+++ vyre/__init__.py
@@ -1,3 +1,3 @@
 __app_name__ = "Vyre"
-__version__ = "1.1.4"
+__version__ = "1.1.5"
 __description__ = "Roblox Alt Account Manager"