+++ vyre/ui/browser.py
@@ -236,3 +236,11 @@ class BrowserPanel(QWidget):
 
     def _go_home(self) -> None:
         self._view.setUrl(QUrl(roblox.HOME_URL))
+
+    def paintEvent(self, event):
+        from PySide6.QtGui import QPainter
+        from PySide6.QtWidgets import QStyle, QStyleOption
+        opt = QStyleOption()
+        opt.initFrom(self)
+        p = QPainter(self)
+        self.style().drawPrimitive(QStyle.PE_Widget, opt, p, self)